move backend classes to backend package
