+++ docs/apidocs/ibmq-provider.rst
@@ -14,3 +14,4 @@ Qiskit IBM Quantum Provider API Reference
    ibmq_visualization
    ibmq_jupyter
    ibmq_utils
+   ibmq_backend
