+++ docs/apidocs/ibmq_backend.rst
@@ -1,0 +1,6 @@
+.. _qiskit-providers-ibmq-backend:
+
+.. automodule:: qiskit.providers.ibmq.backend
+   :no-members:
+   :no-inherited-members:
+   :no-special-members:
