+++ qiskit/providers/ibmq/backend/ibmqbackend.py
@@ -28,21 +28,20 @@ from qiskit.providers.models import (BackendStatus, BackendProperties,
 from qiskit.tools.events.pubsub import Publisher
 from qiskit.providers.models import (QasmBackendConfiguration,
                                      PulseBackendConfiguration)
-
 from qiskit.providers.ibmq import accountprovider  # pylint: disable=unused-import
-from .apiconstants import ApiJobShareLevel, ApiJobStatus, API_JOB_FINAL_STATES
-from .job.utils import api_status_to_job_status
-from .api.clients import AccountClient
-from .api.exceptions import ApiError
+
 from .backendjoblimit import BackendJobLimit
-from .credentials import Credentials
-from .exceptions import (IBMQBackendError, IBMQBackendValueError,
-                         IBMQBackendApiError, IBMQBackendApiProtocolError,
-                         IBMQBackendJobLimitError)
-from .job import IBMQJob
-from .utils import update_qobj_config, validate_job_tags
-from .utils.converters import utc_to_local_all, local_to_utc
-from .utils.json_decoder import decode_pulse_defaults, decode_backend_properties
+from .exceptions import (IBMQBackendError, IBMQBackendValueError, IBMQBackendApiError,
+                         IBMQBackendApiProtocolError, IBMQBackendJobLimitError)
+from ..apiconstants import ApiJobShareLevel, ApiJobStatus, API_JOB_FINAL_STATES
+from ..job import IBMQJob
+from ..job.utils import api_status_to_job_status
+from ..api.clients import AccountClient
+from ..api.exceptions import ApiError
+from ..credentials import Credentials
+from ..utils import update_qobj_config, validate_job_tags
+from ..utils.converters import utc_to_local_all, local_to_utc
+from ..utils.json_decoder import decode_pulse_defaults, decode_backend_properties
 
 logger = logging.getLogger(__name__)
