+++ qiskit/providers/ibmq/jupyter/config_widget.py
@@ -20,7 +20,7 @@ from typing import Union
 import ipywidgets as wid
 from qiskit.test.mock.fake_backend import FakeBackend
 from qiskit.providers.ibmq.visualization.interactive import iplot_gate_map
-from qiskit.providers.ibmq.ibmqbackend import IBMQBackend
+from qiskit.providers.ibmq.backend.ibmqbackend import IBMQBackend
 
 
 def config_tab(backend: Union[IBMQBackend, FakeBackend]) -> wid.GridBox:
