+++ qiskit/providers/ibmq/visualization/interactive/gate_map.py
@@ -17,7 +17,7 @@
 from typing import Tuple, Union, Optional, List
 
 import plotly.graph_objects as go
-from qiskit.providers.ibmq.ibmqbackend import IBMQBackend
+from qiskit.providers.ibmq.backend.ibmqbackend import IBMQBackend
 
 from .plotly_wrapper import PlotlyWidget, PlotlyFigure
 from ..device_layouts import DEVICE_LAYOUTS
