+++ setup.py
@@ -73,6 +73,7 @@ setup(
               'qiskit.providers.ibmq.api.clients',
               'qiskit.providers.ibmq.api.rest',
               'qiskit.providers.ibmq.api.rest.utils',
+              'qiskit.providers.ibmq.backend',
               'qiskit.providers.ibmq.credentials',
               'qiskit.providers.ibmq.job',
               'qiskit.providers.ibmq.managed',
