+++ test/ibmq/test_ibmq_provider.py
@@ -19,7 +19,7 @@ from datetime import datetime
 from qiskit import ClassicalRegister, QuantumCircuit, QuantumRegister
 from qiskit.providers.exceptions import QiskitBackendNotFoundError
 from qiskit.providers.ibmq.accountprovider import AccountProvider
-from qiskit.providers.ibmq.ibmqbackend import IBMQSimulator, IBMQBackend
+from qiskit.providers.ibmq.backend.ibmqbackend import IBMQSimulator, IBMQBackend
 from qiskit.qobj import QobjHeader
 from qiskit.test import providers, slow_test
 from qiskit.compiler import assemble, transpile
